separate functionality into modules and implement main module
