+++ tictactoegui.py
@@ -1,174 +1,8 @@
 from PyQt5.QtWidgets import QMainWindow, QApplication, QPushButton, QMessageBox, QComboBox, QLCDNumber
 from PyQt5.QtCore import Qt
 from PyQt5 import uic
-import sys
-import random
-import heapq
-from collections import deque
 
-
-# Función DFS
-def DFS(graph, visited, stack):
-
-    '''
-    graph -> dictionary of the graph
-    visited -> set of visited nodes
-    '''
-
-    while stack:
-        node = stack.pop()
-        if node not in visited:
-            stack.extend(graph[node])
-            return node
-
-# Función BFS
-def BFS(graph, visited, queue):
-
-    '''
-    graph -> dictionary of the graph
-    visited -> set of visited nodes
-    '''
-
-    while queue:
-        node = queue.popleft()  # Dequeue a node from front of queue
-        if node not in visited:
-            queue.extend(graph[node])  # Enqueue all neighbours
-            return node
-
-# Heurística
-def heuristic(board, position, usr_mark, oponent_mark):
-    #print(position)
-    score = 0
-    win_paths = [
-        [(position[0], 0), (position[0], 1), (position[0], 2)],  # Todas las filas
-        [(0, position[1]), (1, position[1]), (2, position[1])]   # Todas las columnas
-    ]
-
-    # Añadir las diagonales si aplica
-    if position in [(0, 0), (1, 1), (2, 2)]:
-        win_paths.append([(0, 0), (1, 1), (2, 2)])  # Diagonal principal
-    if position in [(0, 2), (1, 1), (2, 0)]:
-        win_paths.append([(0, 2), (1, 1), (2, 0)])  # Diagonal secundaria
-
-    #print(win_paths)
-    # Evaluar cada camino de victoria
-    for path in win_paths:
-        path_symbols = [board[row][col] for row, col in path]
-        #print(path_symbols)
-        #print()
-        if path_symbols.count(oponent_mark) == 2 and usr_mark not in path_symbols:
-            score += 1000  # 2 símbolos del oponente juntos
-        elif path_symbols.count(usr_mark) == 2 and oponent_mark not in path_symbols:
-            score += 100  # 2 símbolos del usuario juntos
-        elif path_symbols.count(usr_mark) == 1 and path_symbols.count(oponent_mark) == 0:
-            score += 10  # 1 símbolo del usuario
-        elif path_symbols.count(oponent_mark) == 1 and path_symbols.count(usr_mark) == 0:
-            score += 1  # 1 símbolo del oponente
-    return score
-
-# Función GBFS
-def GBFS(graph, board, visited, usr_mark, oponent_mark):
-    priority_queue = []
-
-    for node in graph:
-        if node not in visited:
-            score = heuristic(board, node, usr_mark, oponent_mark)
-            heapq.heappush(priority_queue, (-score, node))  # Prioridad por el puntaje más alto
-
-    if sum([_[0] for _ in priority_queue]) == 0 and len(visited) == 0:
-        visited.add((1,1))
-        return (1, 1)
-
-    while priority_queue:
-        # Desencolar el nodo con el puntaje más alto
-        current_score, current_node = heapq.heappop(priority_queue)
-        position = current_node
-        
-        #print(position, visited)
-        if position not in visited:
-            visited.add(position)
-            return position
-    return None
-
-class Board:
-    def __init__(self, usr_mark='X', oponent_mode='Random'):
-        self.idle_char = 'I'
-        self.board = [[self.idle_char for j in range(3)] for i in range(3)]
-        self.graph = {(0, 0): [(0, 1), (1, 1), (1, 0)],
-                      (0, 1): [(0, 0), (1, 0), (1, 1), (1, 2), (0, 2)],
-                      (0, 2): [(0, 1), (1, 1), (1, 1)],
-                      (1, 0): [(0, 0), (0, 1), (1, 1), (2, 1), (2, 0)],
-                      (1, 1): [(0, 0), (0, 1), (0, 1), (1, 0), (1, 2), (2, 0), (2, 1), (2, 2)],
-                      (1, 2): [(0, 2), (0, 1), (1, 1), (2, 1), (2, 2)],
-                      (2, 0): [(1, 0), (1, 1), (2, 1)],
-                      (2, 1): [(2, 0), (1, 0), (1, 1), (1, 2), (2, 2)],
-                      (2, 2): [(2, 1), (1, 1), (1, 2)]}
-        self.available_positions = {(i, j) for i in range(3) for j in range(3)}
-        self.not_available_positions = set()
-        self.usr_mark = usr_mark
-        self.oponent_mark = 'O' if usr_mark == 'X' else 'X'
-
-        self.oponent_mode = oponent_mode
-        self.start_position = (random.choice(list(self.available_positions)))
-
-        self.stack = [self.start_position]
-        #self.stack = [(0,0)]
-        self.queue = deque([self.start_position])
-        #self.queue = deque([(0,0)])
-
-    # Método para checkear el resultado
-    def result_check(self):
-        if self.board[0][0] == self.board[0][1] == self.board[0][2] != self.idle_char:
-            return True
-        elif self.board[1][0] == self.board[1][1] == self.board[1][2] != self.idle_char:
-            return True
-        elif self.board[2][0] == self.board[2][1] == self.board[2][2] != self.idle_char:
-            return True
-        elif self.board[0][0] == self.board[1][0] == self.board[2][0] != self.idle_char:
-            return True
-        elif self.board[0][1] == self.board[1][1] == self.board[2][1] != self.idle_char:
-            return True
-        elif self.board[0][2] == self.board[1][2] == self.board[2][2] != self.idle_char:
-            return True
-        elif self.board[0][0] == self.board[1][1] == self.board[2][2] != self.idle_char:
-            return True
-        elif self.board[0][2] == self.board[1][1] == self.board[2][0] != self.idle_char:
-            return True
-        else:
-            return False
-
-    # Método para manejar el movimiento del usuario
-    def user_move(self, position):
-        if position in self.available_positions:
-            self.board[position[0]][position[1]] = self.usr_mark
-            self.available_positions.remove(position)
-            self.not_available_positions.add(position)
-
-    # Método para manejar el movimiento del oponente
-    def oponent_move(self):
-        if self.oponent_mode == 'Random':
-            oponent_position = random.choice(list(self.available_positions))
-        elif self.oponent_mode == 'DFS':
-            oponent_position = DFS(self.graph, self.not_available_positions, self.stack)
-            if oponent_position == None:
-                self.stack.extend(self.graph[self.start_position])
-                oponent_position = DFS(self.graph, self.not_available_positions, self.stack)
-        elif self.oponent_mode == 'BFS':
-            oponent_position = BFS(self.graph, self.not_available_positions, self.queue)
-            if oponent_position == None:
-                self.queue.extend(self.graph[self.start_position])
-        elif self.oponent_mode == 'GBFS':
-            oponent_position = GBFS(self.graph, self.board, self.not_available_positions, self.usr_mark, self.oponent_mark)
-            if oponent_position == None:
-                self.queue.extend(self.graph[self.start_position])
-        
-        self.board[oponent_position[0]][oponent_position[1]] = self.oponent_mark
-        self.available_positions.remove(oponent_position)
-        self.not_available_positions.add(oponent_position)
-
-        return oponent_position
-
-
+from board import Board
 
 class TicTacToeGui(QMainWindow):
     def __init__(self):
@@ -353,10 +187,3 @@ class TicTacToeGui(QMainWindow):
         self.comboMark.setEnabled(True)
         self.finished = False
         self.setFocus()
-
-
-if __name__ == '__main__':
-    app = QApplication(sys.argv)
-    TicTacToeWindow = TicTacToeGui()
-    app.exec_()
-
